Move transform class T into module (for more readable code)
